requirements.txt
bposd>=1.6
ldpc==0.1.60
matplotlib>=3.8.3
numpy>=1.24.3,<2.0
scipy>=1.13.0
pandas
pyyaml
relay_bp
black